backend/requirements.txt
chainlit==1.3.2
llama-index-vector-stores-azureaisearch==0.3.0
llama-index-core==0.12.2
azure-storage-blob==12.24.0
python-dotenv==1.0.1
llama-index==0.12.2
matplotlib==3.9.3
llama-index-multi-modal-llms-azure-openai==0.3.0
llama-index-llms-azure-openai==0.3.0
llama-index-embeddings-azure-openai==0.3.0
streamlit==1.40.2
IPython==8.30.0
Flask==3.1.0
flasgger==0.9.7.1
sentence-transformers==3.3.1